requirements.txt
PyYAML~=5.3.1
Appium-Python-Client~=1.0.2
selenium~=3.141.0
behave~=1.2.6
allure-behave~=2.8.18
